pyproject.toml
[build-system]
requires = ["setuptools >= 35.0.2", "wheel >= 0.29.0", "poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
force_grid_wrap = 0
force_sort_within_sections = true
forced_separate = "tests"
include_trailing_comma = true
indent = "    "
known_first_party = "examples,script,simplipy,tests"
line_length = 88
multi_line_output = 3
not_skip = "__init__.py"
sections = "FUTURE,STDLIB,INBETWEENS,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
use_parentheses = true

[tool.poetry]
name = "simplisafe-python"
version = "7.3.0"
description = "A Python3, async interface to the SimpliSafe API"
readme = "README.md"
authors = ["Aaron Bach <bachya1208@gmail.com>"]
license = "MIT"
repository = "https://github.com/bachya/simplisafe-python"
packages = [
    { include = "simplipy" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[tool.poetry.dependencies]
aiohttp = "^3.6.2"
python = "^3.7.0"
python-engineio = ">= 3.9.3"
python-socketio = ">=4.3.1"
pytz = "^2019.3"
voluptuous = "^0.11.7"
websockets = "^8.1"

[tool.poetry.dev-dependencies]
Sphinx = "^2.2.1"
aresponses = "^1.1.1"
pre-commit = "^2.0.1"
pytest = "^5.3.5"
pytest-aiohttp = "^0.3.0"
pytest-cov = "^2.8.1"
